clean up modules
